+++ src/test/SingletonTest.java
@@ -1,0 +1,15 @@
+import org.junit.jupiter.api.Test;
+import 클래스.Singleton;
+
+public class SingletonTest {
+
+    @Test
+    public void 싱글톤(){
+        Singleton s = Singleton.getInstance();
+        Singleton ss = Singleton.getInstance();
+
+        System.out.println(s);
+        System.out.println(ss);
+    }
+
+}
